Cleanup and fixes
Cleaned up a lot of unneeded cruft/old code/unnecessary comments.

Added javadocs in a number of classes/methods to help make them easier
to understand.

Added commented out code to support the 3rd peg detector.

Added code to try a different way of doing the turn to heading (old code
is still there, just in different methods named old____() )

Adjusted backup time in all side auto modes

Adjusted code in the Robot.java for the LED Ring while in teleop mode.

Changed the Dashboard output for the Gear Detector to be a Boolean so it
can be a nice big indicator on the dash.

Changed a few things in the Camera system to try to remove the issue
with the switching and streaming problems.

Added a smartdashboard output to indicate if the peg is detected as a
boolean so we can do it as a nice big output for the driver.